♻️ Refactor: Modal의 형식을 분리하고, switch구문을 통해 해당 컴포넌트를 불러오는 식으로 관리하게 함
